pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "airo-drake"
version = "0.0.1"
authors = [
  { name = "Andreas Verleysen", email = "andreas.verleysen@ugent.be" },
  { name = "Victor-Louis De Gusseme", email = "victorlouisdg@gmail.com" },
]
description = "Integration between airo-mono and Drake."
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]

dependencies = [
    "airo-models",
    "drake",
]

[project.urls]
Homepage = "https://github.com/airo-ugent/airo-drake"
Issues = "https://github.com/airo-ugent/airo-drake/issues"
